requirements.txt
pymongo
fastapi
openai
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
chromadb
googlesearch-python
trafilatura
lxml-html-clean